Download API definition from S3 instead of API Gateway
This (at least temporary) workaround addresses the issue
where API Gateway-exported definitions would lose the type
of query string parameters, e.g. 'integer' would be exported as 'string'.

This unfortunate non-compliance to strict OpenAPI means
we have to import the initial definitions if we'd like to preserve the type.

In the future, we expect to publish this definition through CloudFront
and then import it manually (it is not much more work
than specifying API or bucket name on the command line).