Include compiled Codegen 2.3 for the latest features for Angular 2
This change uses the improvements for Angular 2 client generation
in Codegen 2.3, in particular the naming of services according to
Angular style guide.

We're also installing JRE to run Codegen on CodeBuild.
(We opt to re-generate the client on every build to make sure
our it stays current with the upstream definition of the API)