Set up ApiModule from CoreModule
We now import ApiModule into CoreModule and configure it
with the values expected at runtime.

Namely, we pass in the API root URL, as well as
Authorization token from localStorage, and
configure it for use with CORS.

We also test that AppComponent initializes all
of the services from ApiModule when it starts.