Reduce the number of datasets in a page
We notice that most modern screens have effective vertical
resolution up to ~1400 (2880 with 200% upscaling
on an iMac 5K). Since each row is 50px, we only need to
load up to ~28 datasets to fit on the screen. We set it to 30
such that the scroll bar (and thus paging on scroll)
is always available, unless there are fewer datasets to be loaded.