Implement a shared list component with scrolling
We create a base component for displaying any of our datatable
lists using the (scroll) event from NgxDatatable.

Using (scroll) is more precise than (page), because now
we can directly detect an 'end of scrolling' event
and load only as many datasets as defined through pageLimit,
irregardless of how many rows can fit on the screen.

This component is also designed to be general enough
such that the derived classes require only minimal setup
with specifics unique to each class.