Fix page limit calculation for the initial load in ListBaseComponent
Initially (when there are not items in the list),
we must load not less than the number of items that can fit in the view
(otherwise, the scrollbar will not appear, and we won't be able to scroll).

Afterwards, the page limit is simply the value set for the class.