Export dynamic Authorization key object in CoreModule
We no longer define an initial value for Authorization key.

Instead, we export the apiKeys object and pass it
as part of configuration of ApiModule.

This approach enables modification of the key at runtime
(to be later set in AuthService).

We also move tests for ApiModule config from AppComponent to CoreModule.