Move list loading indicator to the bottom only after the initial load.
Previously, we always displayed the progress indicator at the bottom
of the page, which made it less visible during the initial load.

Now, users can more easily see that the data is still been loaded
when a component list is initialized.

Afterwards, it will always be displayed at the bottom to indicate
the "direction" of loading of additional results.