Upgrade Ngx-Datatable and related packages
The new version fixes a few UI issues for row details.
It also requires a newer version of @angular packages.

Additionally, it updates internal styles so we had to adjust the font size.

Finally, it now also invokes 'activate' event on 'mouseover',
so we had to filter it to only react to clicks.