Show item details in a dialog instead of 'inline' in ListBaseComponent
We now use MdDialogModule to inject a dialog element into
the list component, and open it when the user clicks on a list item.

This process simplifies rendering and editing of item details,
because it 'focuses' on one item at a time, instead of
'goofy' navigation across multiple forms.

It also removes the need to manage form height explicitly,
because MdDialog will adjust that for us automatically.

Finally, it avoids visual glitches associated with
inline item detail display that we have seen in the past.

TODO: update Project details accordingly.