Define DialogBaseComponent for common functionality of all dialogs
We define and implement DialogBaseComponent with general functionality that
can be extended by application dialogs according to their specific needs.

In particular, we define createForm() method
and use it to instantiate the dialog form in the base constructor.

Later, we can reference this form in the derived class
through the form() accessor.

We also define the update() method to submit form changes to the backend.

On a successful update(), the dialog is closed and
the updated values are passed back to the caller component.

On an unsuccessful update(), the list of 'body' errors from the server response
(if any) is parsed into a dictionary of errors for the corresponding
controls, with an error status set for each control.