Implement a base service for caching items
We create CachingBaseService as a base class to
cache items such as projects, users and clients.

Services derived from this class will be able to
to update and return cached versions of their items
to reduce the amount of requests to the backend.

For example, we will be able to display cached versions
of project names for each item in a list of datasets,
or to display a list of project names without contacting the backend.

This functionality comes with a usual caveat that
we need to explicitly update the cache when a new request
does return updated information.

We may also implement time-based cache expiration in the future.