Implement ProjectNamePipe for displaying project names
We implement a pipe that uses ProjectsCachingService to
display the project name instead of project id.

If the project name is undefined (e.g. for a project that
was not found in the cache), we fallback to displaying the id.

It also uses the AsyncPipe, which should update the name
automatically if the project name changes in the cache.