Implement an initial set of action factories
We implement Store action factories for the initial set of actions.

The action names are prefixed with the name of the 'family' of items,
i.e. projects, datasets, ...

We expect all item families to use homogeneous actions, thus
it is useful to unify these and generate actions through factories
differing only in the name of each family.

For each family, we also use typescript-fsa subfactories to
generate concrete actions.