Create some initial definitions for state management
We define ItemsState class to keep track of collections of items.

We extend this interface as EntityState from @ngrx/entity,
which is a convenience interface for managing items
as both a dictionary and an ordered array.

In addition, the global state will keep track of 'isLoading'
property, which we will use to indicate server interactions in the UI.

We also define a Property class to associate item properties
with their display names, types (e.g. 'date') and whether
they are displayed as columns in the associated lists
(otherwise, they can only be seen through 'details').

This aproach will ensure consistency in displaying various
properties across the entire app, and make visual elements
such as lists and dialogs more generic (i.e. their visual
display will depend entirely on the input config).

It will also let us expand functionality over time,
e.g. to turn on/off individual columns or use them in sorting.