Create reducer factory
We implement a factory to generate a reducer
for a given family of items.

The factory also takes an array of properties to initialize
the initial state.

Currently, we only use actions defined
for the collection of items, while
the properties remain static.

Our implementation uses a convenience 'entity adapter'
to simplify management of @ngrx entities.

Later on, we will modify these methods to properly handle
item caching and other enhancements.