Implement StoreService to wrap interactions with the Store
We create an intermediary to the store in the form of StoreService.

This service proxies requests to the store such that its clients
don't have to deal with the low-level details.

This approach decouples clients from a particular implementation
of the store (which could be switched from NgRx in the future if needed).
It also simplifies the overall syntax.

Currently, we implement only select() and list() methods,
which will be immediately useful to the List* components.