Define EffectsService for interaction between the store and the API
We define an abstract EffectsService class that can be used
to implement API interactions triggered by actions on the store.

The API results trigger "success" actions with an associated payload.

For now, we only define a "happy" path for each action,
and will need to implement error handling down the road.

We will also add more methods, such as for fetching individual items.