Implement ListViewComponent
We create a dedicated component to display list views.

This component exposes a number of inputs and only deals
with rendering the data, as well as emitting appropriate
events upon user interaction - namely, (list) and (open).

This architecture de-couples list display from the rest
of the logic needed to obtain the data (e.g. through the Store).

In turn, that enables us to reuse this component to
display *any* list, i.e. we no longer have to extend it for
projects, datasets, etc.

Actual logic to supply the data and handle user events
will be provided by an associated "container" component
as a parent of this view.