Implement generalized ListComponent
We create a general-purpose "controller" component that
selects appropriate observables from StoreService
and supplies their values to ListViewComponent.

The new component can be reused to display any list,
we only need to provide items family through its "type" attribute.

Note that this component is solely in charge of supplying data
to the view. It also bubbles up (open) events from the view,
so the task of handling these events is up to the parent
of ListComponent.

This approach provides additional flexibility in
displaying and interacting with the list view.