Implement ItemsPageComponent for displaying list with a toolbar
We create a generic component to display the "default" page
that a user sees when they navigate to an items tab.

This component is meant to replace the item type-specific
pages we implemented earlier.

Since this is the "root" component for item modules,
we pass the "type" through router data instead of an @Input.

Next, we will update our item modules to use this component instead.