Use Router navigation instead of StoreService actions (incl. ListComponent)
We remove StoreService and communicate with the store directly
for selections (which have identical API), as well as
with the router instead of the store for actions.

This approach makes StoreModule leaner by removing unnecessary
indirection for communication with the store.

In addition, store actions are dispatched through the router,
which acts as the source of truth.

ListComponent is the only client that dispatched actions
before this change.