Sync Store with localStorage
We use ngrx-storage-localstorage to automatically
save application state to localStorage,
as well as to rehydrate it on app init.

We also implement syncing between tabs,
with the help of StorageEvents that get dispatched
when localStorage is modified in one of the tabs
(and we consider storage to be the source of truth).

In this case, other tabs update their state in memory
without committing the new state to storage again
(otherwise, we'd get an infinite loop).

Currently, we store only the collections, whereas
router state is considered transient for each tab.