Re-implement AuthModule to use the Store state
We refactor AuthService with the use of
the authentication state from the store.

This approach enables us to simplify the service,
caring only about selection and transformation
of state, while delegating login() and logout()
functions to derived provider classes.

As such, we no longer need a separate AuthProviderService,
rather we derive CognitoAuthService with
pure functions that construct CognitoAuth object
and redirect to provider urls.

These functions will only be called from
associated effects in response to
LOGIN_REDIRECT and LOGOUT actions.

This approach enables very succinct
implementations of authentication services,
because we no longer need to manage app state.

In fact, we no longer use CognitoAuth's
internal storage mechanism for its tokens,
which helps us avoid a few subtleties
associated with its behavior.

We also refine AuthGuardService to get
authentication status from AuthService
and navigate to a /login endpoint if needed.

Additionally, we remove canActivateChild
and canLoad actions as unnecessary
(we allow loading in the background
to speed up the app).