Update Login and Session components to interact solely with the router
Since we now treat the router as the source of truth,
we no longer need (and should not) directly call
login() and logout() of AuthService.

Instead, we either navigate to logout URL in SessionComponent,
or simply render a (for now) empty page in LoginComponent
(the actual logic of login redirect and logout happens in AuthEffects).