Create store state for the API
We add API settings (currently, only baseUrl) to the store
to complete the transition to reactive state architecture.

However, we still need to re-design ApiModule
to use this state (along with the auth state).

Technically, baseUrl is immutable, so we could have stored
it outside of the store. However, including it here helps
us make all state access homogenenous and
enables its visualization in DevTools,
as well as to troubleshoot violations of immutability.

After these changes, our entire app will become
a function of the store, with no persistent or client state
stored in services and components (we may still leverage
transient/UI state separately, however).