Create ApiModule with models
We replace auto-generated ApiModule with
one that will make use of the store.

Currently, we only define models for
projects, datasets and analyses
and update localStorage revival function to
never revive dates (which was "on" by default).

We also remove code generation from api-client.sh,
and delete the associated CodeGen binary.

Finally, we temporarily remove SwaggerComponent,
which relied on the old ApiModule.