Use a dedicated testing environment for unit tests
Previously, we relied on environment config that
used real values for the beta and production backend stacks.

For unit tests, that is unnecessary and actually
fails tests in CircleCI because we don't set up
these environments for public builds.

Now, we define a dedicated 'test' environment
with fake values that mimic real ones but
don't relate to the actual stacks.

This change enables us to fully de-couple unit tests
from the backend deployments.