Use ApiService in EffectsService
Previously, we deferred implementation of create(), update()
and list() methods to the derivatives of EffectsService.

Now, we call ApiService directly in a generic way,
such that derived classes only need to register the effects
from the superclass and supply the type of items.

This will enable us to significantly cut down the amount
of boilerplate code and testing for concrete implementations
of effects.