Fetch projects referenced in item collections
So far, all collections fetched in response to LIST action
would be displayed with a cryptic 'project_id'.

Now, we add an effect for LIST_SUCCESS action
that outputs GET action for each project_id
this is not yet in the store
(to be displayed later as e.g. the project name).

These actions happen asynchronously after LIST_SUCCESS,
which will need additional indication in the UI.