Display project names instead of IDs in List*Component
Previously, project IDs would be displayed in the "raw" form,
which is not particularly human-friendly.

Now, we select projects from the store and display
their names instead of IDs in the lists.

If an ID is not found, we skip its display
and rely on background fetching of project
details through store effects.