Move custom list cell display into a dedicated component
We now display list project names via
ListCellComponent, which shows
its input item name (or a loading indicator)
irregardless of what type of item it is.

This change will enable us to not only easily display
project (and later user/client/...) names,
but also keep this logic implemented in a single place,
with extensibility for displaying errors etc.