Require generic Item for Store classes to have a string id
Previously, we did define some items as StoreItem
that had an "id" of type "any".

However, all of our data models have a string "id" so
it makes sense to narrow that down.

Additionally, we now enforce generic Item
to be extended from StoreItem everywhere,
which ensures consistency and provides
additional protection with the type check.