Update compiler configs for Angular 5
We put tsconfig and related files more in line
with those auto-generated for a new Angular 5 project.

This change avoids duplication of compiler options,
making these configurations leaner and more to the point.

We also fix name for Angular environment in Karma
and a minor compiler-related issue in AuthService spec
(which also appeared to be unused code).