Implement request interceptor for the API
We register an interceptor that appends
Authorization header and baseUrl to each request.

This interceptor together with HttpClient will be
used in place of ApiService later on.

This approach frees us from having to wrap
API methods for each type of request (GET, POST, ...).

Instead, the HttpClient can be injected and mocked
directly into dependent services.