Use HttpClient in Effects and remove ApiService
We replace calls to ApiService with those for HttpClient.

This approach results in smaller code,
obviating the need for any wrappers for HttpClient.

We also fix UPDATE effect by producing UPDATE_SUCCESS
action with the corresponding HTTP response
instead of the initial payload.

Finally, we use object destructuring
to cut down the code in EffectsService.