Add reducer for FAILURE action factories
We expand definition of StoreItem to optionally
be an Error object.

This approach aligns with FSA guidelines for
passing Error as the payload.

We store this error for GET_FAILURE action,
and ignore it (by only reducing requestCount)
in other failure modes.

This approach will enable display of item-level
errors directly in list views.