Move requirement for no unused variables to TSLint config
Recently, we introduced noUnusedLocals and noUnusedParameters
to enforce removal of unused variables from the code.

However, this was a compile-time check that turned out
more annoying than necessary in development.

Now, we offload this check to the linter, so
while it is still enforced in CI and pre-commit hook,
we do have more freedom during the dev cycle
between commits.

As a side effect, that also revealed unused imports in e2e.